Abort db migration if data_directory is defined in postgresql conf
The migration script assumes data_directory is not defined. Otherwise, it
would give an error.

This commit prevents the error by checking and aborting it.

+++ susemanager/bin/pg-migrate-10-to-12.sh
@@ -52,6 +52,12 @@ else
     BACKUP_CONFIGURED=0
 fi
 
+if [ $(grep data_directory /var/lib/pgsql/data/postgresql.conf) ]; then
+    echo "$(timestamp) data_directory is configured in /var/lib/pgsql/data/postgresql.conf"
+    echo "$(timestamp) For the migration to work, data_directory should not be defined explicetely"
+    exit 1
+fi
+
 echo "`date +"%H:%M:%S"`   Shut down spacewalk services..."
 spacewalk-service stop
 systemctl stop postgresql
